trocr_requirements.txt
numpy<2
pillow 
opencv-python
transformers
torch 
torchvision